+++ src/components/AudioPlayer/RepeatAudioModalProps.tsx
@@ -1,0 +1,121 @@
+import { useMemo, useState, useEffect } from 'react';
+
+import { RangeSelectorType, RangeVerseItem } from '../Verse/AdvancedCopy/SelectorContainer';
+
+import styles from './RepeatButton.module.scss';
+
+import Counter from 'src/components/dls/Counter/Counter';
+import Modal from 'src/components/dls/Modal/Modal';
+import VerseRangeSelector from 'src/components/Verse/AdvancedCopy/VersesRangeSelector';
+import { getChapterData } from 'src/utils/chapter';
+import { generateChapterVersesKeys } from 'src/utils/verse';
+
+type RepeatAudioModalProps = {
+  chapterId: string;
+  isOpen: boolean;
+  onClickOutside: () => void;
+};
+
+const RepeatAudioModal = ({ chapterId, isOpen, onClickOutside }: RepeatAudioModalProps) => {
+  const chapterName = useMemo(() => {
+    const chapterData = getChapterData(chapterId);
+    return chapterData?.nameSimple;
+  }, [chapterId]);
+
+  const rangeVersesItems = useMemo<RangeVerseItem[]>(() => {
+    const keys = generateChapterVersesKeys(chapterId);
+
+    const initialState = keys.map((chapterVersesKey) => ({
+      id: chapterVersesKey,
+      name: chapterVersesKey,
+      value: chapterVersesKey,
+      label: chapterVersesKey,
+    }));
+
+    return initialState;
+  }, [chapterId]);
+
+  // eslint-disable-next-line prefer-destructuring
+  const firstVersesRangeItems = rangeVersesItems[0];
+  const lastVersesRangeItems = rangeVersesItems[rangeVersesItems.length - 1];
+
+  // TODO: connect to redux when the data flow is ready
+  const [repeatVerse, setRepeatVerse] = useState({ total: 1, progress: 0 });
+  const [repeatVerseRange, setRepeatVerseRange] = useState({
+    total: 1,
+    progress: 0,
+    from: firstVersesRangeItems.value, // first verseKey in the current chapter
+    to: lastVersesRangeItems.value, // last verseKey in the current chapter
+  });
+
+  // reset repeatVerseRange's `to` and `from`, when chapter changed
+  useEffect(() => {
+    setRepeatVerseRange((prevRepeatVerseRange) => ({
+      ...prevRepeatVerseRange,
+      from: firstVersesRangeItems.value,
+      to: lastVersesRangeItems.value,
+    }));
+  }, [chapterId, firstVersesRangeItems.value, lastVersesRangeItems.value]);
+
+  return (
+    <Modal isOpen={isOpen} onClickOutside={onClickOutside}>
+      <Modal.Body>
+        <Modal.Header>
+          <Modal.Title>Repeat Settings</Modal.Title>
+          <Modal.Subtitle>Surah {chapterName}</Modal.Subtitle>
+        </Modal.Header>
+        <div>
+          <div className={styles.inputContainer}>
+            <VerseRangeSelector
+              onChange={(value, dropdownId) => {
+                if (dropdownId === RangeSelectorType.END) {
+                  setRepeatVerseRange({ ...repeatVerseRange, to: value });
+                } else {
+                  setRepeatVerseRange({ ...repeatVerseRange, from: value });
+                }
+              }}
+              dropdownItems={rangeVersesItems}
+              isVisible
+              rangeStartVerse={firstVersesRangeItems.value}
+              rangeEndVerse={lastVersesRangeItems.value}
+            />
+          </div>
+          <div className={styles.inputContainer}>
+            <span className={styles.label}>Play each verse:</span>{' '}
+            <span className={styles.input}>
+              <Counter
+                onIncrement={() => {
+                  setRepeatVerse({ ...repeatVerse, total: repeatVerse.total + 1 });
+                }}
+                onDecrement={() => {
+                  if (repeatVerse.total >= 0)
+                    setRepeatVerse({ ...repeatVerse, total: repeatVerse.total - 1 });
+                }}
+                count={repeatVerse.total}
+              />{' '}
+              times
+            </span>
+          </div>
+          <div className={styles.inputContainer}>
+            <span className={styles.label}>Play verse range:</span>{' '}
+            <span className={styles.input}>
+              <Counter
+                onIncrement={() => {
+                  setRepeatVerseRange({ ...repeatVerseRange, total: repeatVerseRange.total + 1 });
+                }}
+                onDecrement={() => {
+                  if (repeatVerseRange.total >= 0)
+                    setRepeatVerseRange({ ...repeatVerseRange, total: repeatVerseRange.total - 1 });
+                }}
+                count={repeatVerseRange.total}
+              />{' '}
+              times
+            </span>
+          </div>
+        </div>
+      </Modal.Body>
+    </Modal>
+  );
+};
+
+export default RepeatAudioModal;
